showroom/usecase-045/requirements.txt
flask>=2.0.0,<3.0.0
pandas>=1.3.0,<2.2.0
numpy>=1.20.0,<2.0.0
plotly>=5.0.0,<6.0.0
openai>=1.0.0
python-dotenv>=0.19.0
flask-wtf>=1.0.0
requests>=2.25.0
matplotlib>=3.5.0
seaborn>=0.11.0